+++ libs/ts-configurable/src/index.ts
@@ -1,3 +1,3 @@
-export { IConstructorOptions } from './lib/interfaces';
+export { IConstructorOptions, IEnvOptions, IDecoratorOptions } from './lib/interfaces';
 export { Configurable } from './lib/configurable';
 export { BaseConfig } from './lib/base-config';