+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.0
 
 - Update documentation
+- Additionally export IDecoratorOptions and IEnvOptions from package
 - Change throwing "Error" to "TypeError"
 
 ## 0.0.2